Clean up the new test code
